+++ app/lib/llmizer_web/live/chats_index_live.ex
@@ -1,22 +1,10 @@
-defmodule LlmizerWeb.ChatsLive do
+defmodule LlmizerWeb.ChatsIndexLive do
   use LlmizerWeb, :live_view
 
   alias Llmizer.Chats
   alias LlmizerWeb.ChatFormComponent
-  alias Llmizer.Chats.ChatMessage
 
   @impl true
-  def mount(%{"chat_id" => chat_id}, _session, socket) do
-    chats = Chats.list_chats()
-    chat = Chats.get_chat!(chat_id)
-
-    {:ok,
-     socket
-     |> assign(:page_title, "#{chat.name}}")
-     |> assign(:chats, chats)
-     |> assign(:chat, chat)}
-  end
-
   def mount(_params, _session, socket) do
     chats = Chats.list_chats()
 
@@ -38,9 +26,6 @@ defmodule LlmizerWeb.ChatsLive do
       <% else %>
         <p>No chats available.</p>
       <% end %>
-      <%= if @chat do %>
-        {chat_messages(assigns)}
-      <% end %>
     </div>
     """
   end
@@ -60,19 +45,6 @@ defmodule LlmizerWeb.ChatsLive do
     """
   end
 
-  defp chat_messages(assigns) do
-    ~H"""
-    <div>
-      <h2>{@chat.name}</h2>
-      <ul>
-        <%= for message <- @chat.chat_messages do %>
-          <li><strong>{message.role}:</strong> {message.content}</li>
-        <% end %>
-      </ul>
-    </div>
-    """
-  end
-
   defp new_chat_message_changeset do
     attrs = %{
       content: "",
@@ -85,6 +57,7 @@ defmodule LlmizerWeb.ChatsLive do
      |> to_form()
   end
 
+  @impl true
   def handle_event("validate", %{"chat_message" => chat_message_params}, socket) do
     changeset = new_chat_message_changeset()
     |> Map.put(:params, chat_message_params)
@@ -93,14 +66,11 @@ defmodule LlmizerWeb.ChatsLive do
   end
 
   def handle_event("save", %{"chat_message" => chat_message_params}, socket) do
-    dbg(chat_message_params)
     case Chats.create_chat_message(chat_message_params) do
       {:ok, _chat_message} ->
-        dbg(_chat_message)
         {:noreply, socket |> put_flash(:info, "Chat message created successfully.")}
 
       {:error, changeset} ->
-        dbg(changeset)
         {:noreply, assign(socket, :form, changeset)}
     end
   end
