+++ app/lib/llmizer_web/router.ex
@@ -17,8 +17,8 @@ defmodule LlmizerWeb.Router do
   scope "/", LlmizerWeb do
     pipe_through :browser
 
-    live "/", ChatsLive, :home
-    live "/chats/:chat_id", ChatsLive, :show
+    live "/", ChatsIndexLive
+    live "/chats/:chat_id", ChatsShowLive
   end
 
   # Other scopes may use custom stacks.
