+++ app/test/llmizer_web/live/chats_show_live_test.exs
@@ -1,25 +1,6 @@
-defmodule LlmizerWeb.ChatsLiveTest do
+defmodule LlmizerWeb.ChatsShowLiveTest do
   use LlmizerWeb.ConnCase
   alias Llmizer.Chats
-
-  describe "GET /index" do
-    test "renders list of chats", %{conn: conn} do
-      Chats.create_chat(%{name: "Test Chat"})
-
-      conn = get(conn, ~p"/")
-      assert html_response(conn, 200) =~ "Your chats"
-      assert html_response(conn, 200) =~ "Test Chat"
-    end
-  end
-
-  describe "POST /chats" do
-    test "creates one chat", %{conn: conn} do
-      conn = post(conn, ~p"/chats", chat_message: %{content: "New Chat"})
-      assert redirected_to(conn) == ~p"/chats/1"
-
-      assert Chats.get_chat!(1).name == "New Chat"
-    end
-  end
 
   describe "GET /chats/:id" do
     test "renders one chat", %{conn: conn} do
